Add normalizer to change all remove/add enrage to refresh.
When enrage is continously "removed" and "added", it can look bad in timeline views and clutters down the history. This normalizer will change all remove/add enrage to single refresh events instead.
Following similiar structure of EventOrder, and Link normalizers, a generic base was added, that implmentations can extend to provide what spell values are to be considered for the normalizer.